xml/parsing/manakai/xmldecls-2.dat

#data escaped
\u0085<p>abc</p>
#errors
1;1;no XML decl;;;s
1;1;control char;;U+0085;w
1;1;text outside of root element;;;m
#document escaped
| <{} p>
|   "abc"

#data escaped
<?xml version="1.1"?>\u0085<p>abc</p>
#errors
1;10;xml:version:not 1.0;;1.1;s
1;22;control char;;U+0085;w
1;22;text outside of root element;;;m
#xml-version
1.1
#document
| <{} p>
|   "abc"

#data escaped
<?xml version="1.1"?>\u000D\u0085<p>abc</p>
#errors
1;10;xml:version:not 1.0;;1.1;s
2;1;control char;;U+0085;w
2;1;text outside of root element;;;m
#xml-version
1.1
#document
| <{} p>
|   "abc"

#data escaped
<?xml version="1.1"?>\u2028<p>abc</p>
#errors
1;10;xml:version:not 1.0;;1.1;s
1;22;text outside of root element;;;m
#xml-version
1.1
#document
| <{} p>
|   "abc"

#data escaped
<?xml version="1.0"?>\u0085<p>abc</p>
#errors
1;22;control char;;U+0085;w
1;22;text outside of root element;;;m
#document escaped
| <{} p>
|   "abc"

#data escaped
<?xml version="1.0"?>\u2028<p>abc</p>
#errors
1;22;text outside of root element;;;m
#document escaped
| <{} p>
|   "abc"

#data escaped
<?xml version="1.1"\u0085?>\u0085<p>abc</p>
#errors
1;10;xml:version:not 1.0;;1.1;s
1;20;bogus XML declaration;;;m
1;20;control char;;U+0085;w
1;23;control char;U+0085;;m
1;23;text outside of root element;;;m
#xml-version
1.1
#document escaped
| <{} p>
|   "abc"

#data escaped
<?xml version="1.1"\u2028?>\u0085<p>abc</p>
#errors
1;10;xml:version:not 1.0;;1.1;s
1;20;bogus XML declaration;;;m
1;23;control char;U+0085;;m
1;23;text outside of root element;;;m
#xml-version
1.1
#document escaped
| <{} p>
|   "abc"
